+++ src/main/kotlin/com/example/ajaxproject/repository/MentesRepository.kt
@@ -6,4 +6,4 @@ import com.example.ajaxproject.model.User
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
-interface MentesRepository : CrudRepository<Mentees, Int>
+interface MentesRepository : CrudRepository<Mentees, Long>
