+++ src/main/kotlin/com/example/ajaxproject/repository/UserRepository.kt
@@ -4,4 +4,4 @@ import com.example.ajaxproject.model.User
 import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
-interface UserRepository : CrudRepository<User, Int>
+interface UserRepository : CrudRepository<User, Long>
